Task3. Example of Abstract Factory (created packages for templates)
