+++ src/main/java/ru/aston/matveenko_ym/patterns/creational/abstract_factory/abstract_products/Payment.java
@@ -1,0 +1,6 @@
+package ru.aston.matveenko_ym.patterns.creational.abstract_factory.abstract_products;
+
+public interface Payment {
+
+    void processPayment(double amount);
+}
