+++ src/main/java/ru/aston/matveenko_ym/patterns/creational/abstract_factory/client/OrderManager.java
@@ -1,8 +1,8 @@
-package ru.aston.matveenko_ym.task3.client;
+package ru.aston.matveenko_ym.patterns.creational.abstract_factory.client;
 
-import ru.aston.matveenko_ym.task3.abstract_factory.OrderFactory;
-import ru.aston.matveenko_ym.task3.abstract_products.Payment;
-import ru.aston.matveenko_ym.task3.abstract_products.Product;
+import ru.aston.matveenko_ym.patterns.creational.abstract_factory.abstract_order_factory.OrderFactory;
+import ru.aston.matveenko_ym.patterns.creational.abstract_factory.abstract_products.Payment;
+import ru.aston.matveenko_ym.patterns.creational.abstract_factory.abstract_products.Product;
 
 public class OrderManager {
     private final Product product;
