--- src/main/java/ru/aston/matveenko_ym/task3/abstract_products/Payment.java
@@ -1,6 +1,0 @@
-package ru.aston.matveenko_ym.task3.abstract_products;
-
-public interface Payment {
-
-    void processPayment(double amount);
-}
