--- src/main/java/ru/aston/matveenko_ym/task3/abstract_products/Product.java
@@ -1,6 +1,0 @@
-package ru.aston.matveenko_ym.task3.abstract_products;
-
-public interface Product {
-
-    void displayInfo();
-}
